remove TopButtonsPanelStrategy.kt
